+++ schema/spacewalk/postgres/procs/sequence_nextval.sql
@@ -1,5 +1,5 @@
-create or replace function sequence_nextval( seq_name regclass ) returns numeric as
+create or replace function
+sequence_nextval( seq_name regclass ) returns bigint as
 $$
-	select nextval( $1 )::numeric;
+	select nextval($1);
 $$ language sql;
-
